restructure context and add tests
